# openapi.yml
openapi: 3.0.3
info:
  title: JihyunReminder API
  description: Apple Reminder 웹 버전 REST API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버

tags:
  - name: ReminderList
    description: 리마인더 목록 관리
  - name: Reminder
    description: 리마인더 관리
  - name: SubTask
    description: 하위 태스크 관리

paths:
  /api/lists:
    get:
      tags:
        - ReminderList
      summary: 전체 목록 조회
      operationId: getAllLists
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReminderListResponse'

    post:
      tags:
        - ReminderList
      summary: 목록 생성
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderListRequest'
            example:
              name: 개인
              color: BLUE
              icon: 🏠
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '400':
          description: 유효하지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/lists/{id}:
    patch:
      tags:
        - ReminderList
      summary: 목록 수정
      operationId: updateList
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderListUpdateRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '404':
          description: 목록 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - ReminderList
      summary: 목록 삭제
      operationId: deleteList
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 삭제 성공
        '404':
          description: 목록 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/reminders:
    get:
      tags:
        - Reminder
      summary: 리마인더 조회 (목록별 / 스마트 목록 / 검색)
      operationId: getReminders
      parameters:
        - name: listId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: 목록 ID로 조회
        - name: smart
          in: query
          required: false
          schema:
            type: string
            enum: [today, scheduled, all, flagged, completed]
          description: 스마트 목록 유형
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: 검색어 (title/memo 대소문자 무시)
      responses:
        '200':
          description: 조회 성공 (미완료/완료 분리)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderGroupResponse'

    post:
      tags:
        - Reminder
      summary: 리마인더 생성
      operationId: createReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderCreateRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '400':
          description: 유효하지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 목록 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/reminders/{id}:
    patch:
      tags:
        - Reminder
      summary: 리마인더 수정
      operationId: updateReminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderUpdateRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: 리마인더 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Reminder
      summary: 리마인더 삭제
      operationId: deleteReminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 삭제 성공
        '404':
          description: 리마인더 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/reminders/{id}/complete:
    patch:
      tags:
        - Reminder
      summary: 리마인더 완료 토글
      operationId: toggleReminderComplete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 토글 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: 리마인더 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/reminders/{id}/flag:
    patch:
      tags:
        - Reminder
      summary: 리마인더 플래그 토글
      operationId: toggleReminderFlag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 토글 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: 리마인더 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/reminders/{id}/move:
    patch:
      tags:
        - Reminder
      summary: 리마인더 목록 이동
      operationId: moveReminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderMoveRequest'
      responses:
        '200':
          description: 이동 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: 리마인더 또는 목록 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/smart-counts:
    get:
      tags:
        - Reminder
      summary: 스마트 목록 카운트 일괄 조회
      operationId: getSmartCounts
      responses:
        '200':
          description: 카운트 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartCountResponse'

  /api/reminders/{reminderId}/subtasks:
    post:
      tags:
        - SubTask
      summary: 하위 태스크 생성
      operationId: createSubTask
      parameters:
        - name: reminderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubTaskRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskResponse'
        '404':
          description: 리마인더 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/subtasks/{id}:
    patch:
      tags:
        - SubTask
      summary: 하위 태스크 제목 수정
      operationId: updateSubTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubTaskRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskResponse'
        '404':
          description: 하위 태스크 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - SubTask
      summary: 하위 태스크 삭제
      operationId: deleteSubTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 삭제 성공
        '404':
          description: 하위 태스크 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/subtasks/{id}/complete:
    patch:
      tags:
        - SubTask
      summary: 하위 태스크 완료 토글
      operationId: toggleSubTaskComplete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 토글 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskResponse'
        '404':
          description: 하위 태스크 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ListColor:
      type: string
      enum:
        - RED
        - ORANGE
        - YELLOW
        - GREEN
        - TEAL
        - BLUE
        - INDIGO
        - PURPLE
        - PINK
        - BROWN
      example: BLUE

    Priority:
      type: string
      enum:
        - NONE
        - LOW
        - MEDIUM
        - HIGH
      example: NONE

    ReminderListRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 50
          example: 개인
        color:
          $ref: '#/components/schemas/ListColor'
        icon:
          type: string
          example: 🏠

    ReminderListUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          example: 업무
        color:
          $ref: '#/components/schemas/ListColor'
        icon:
          type: string
          example: 💼

    ReminderListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: 개인
        color:
          $ref: '#/components/schemas/ListColor'
        icon:
          type: string
          example: 🏠
        createdAt:
          type: string
          format: date-time

    ReminderMoveRequest:
      type: object
      required:
        - listId
      properties:
        listId:
          type: integer
          format: int64
          example: 2

    ReminderCreateRequest:
      type: object
      required:
        - listId
        - title
      properties:
        listId:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          maxLength: 200
          example: 장보기
        memo:
          type: string
          example: 우유, 달걀

    ReminderUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          example: 장보기
        memo:
          type: string
          example: 우유, 달걀
        priority:
          $ref: '#/components/schemas/Priority'
        dueDate:
          type: string
          format: date
          example: 2026-07-01
          nullable: true
        dueTime:
          type: string
          format: time
          example: "09:00:00"
          nullable: true

    SubTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          example: 우유 사기

    SubTaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        reminderId:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: 우유 사기
        completed:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time

    ReminderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        listId:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: 장보기
        memo:
          type: string
          example: 우유, 달걀
        completed:
          type: boolean
          example: false
        completedAt:
          type: string
          format: date-time
          nullable: true
        flagged:
          type: boolean
          example: false
        priority:
          $ref: '#/components/schemas/Priority'
        dueDate:
          type: string
          format: date
          nullable: true
        dueTime:
          type: string
          format: time
          nullable: true
        createdAt:
          type: string
          format: date-time
        subTasks:
          type: array
          items:
            $ref: '#/components/schemas/SubTaskResponse'

    ReminderGroupResponse:
      type: object
      properties:
        incomplete:
          type: array
          items:
            $ref: '#/components/schemas/ReminderResponse'
        completed:
          type: array
          items:
            $ref: '#/components/schemas/ReminderResponse'

    SmartCountResponse:
      type: object
      properties:
        today:
          type: integer
          format: int64
          example: 3
        scheduled:
          type: integer
          format: int64
          example: 10
        all:
          type: integer
          format: int64
          example: 25
        flagged:
          type: integer
          format: int64
          example: 2
        completed:
          type: integer
          format: int64
          example: 15

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: 목록을 찾을 수 없습니다. id=1